Leave kak-lsp server in charge of cleaning up fifo on exit
If we run into a timeout for whatever reason, we remove the fifo and
PID files.  The next request will likely start another kak-lsp server,
which seems like the wrong failure mode.

Let's remove some of the smartness.  If we're in this degraded state,
it's better to fail loud and prompt the user to restart servers
if needed.

Additionally, resetting the fifo option on lsp-exit doesn't seem to
do much, so get rid of it.  We only need it on session rename because
that invalidates the fifo path.