Fix crash when "language" key is used in legacy kak-lsp.toml
Running

	git checkout v14.0.0 -- kak-lsp.toml

we crash on "lsp-diagnostics" because the since 55299017 (Fix multiple
legacy languages w/ same language server clashing, 2024-05-12), when
the legacy "languages" field is used in legacy kak-lsp.toml, we add
the language id to the server name in the map key, to avoid clashes.
However we fail to add the language ID back when looking up server
configs, so do that.

Fixes #792