Don't send buffer contents via fifo
For some reason, when sending over files >1MB,
kak-lsp doesn't manage to read the fifo,
so some requests time out.

Let's work around this by using a separate file.

Not sure if the fifo approach works out;
if we continue to hit problems we should
switch back to a Unix socket.