Work around rust-analyzer crash when requesting code actions for the whole line
When the selections don't contain a code action, we fall back to
requesting code actions for the whole line. This may be unexpected
but it is our way of making up for the lack of per-line lightbulbs.

Instead of computing the actual line end, use EOL_OFFSET=1e6.
This makes rust-analyzer panic with "Bad offset: range 0..97493 offset
1017440" in certain scenarios.  LSP says on Position::character:

> /// If the character value is greater than the line length it
> /// defaults back to the line length.

Work around this until we can expect this to be fixed.

Ref https://github.com/rust-lang/rust-analyzer/issues/18240