Remove extra warnings when a server does not support a request
In the places where this matters (e.g. when the set of eligible
servers is empty), the caller should print a real error.  In other
cases (like when the server does not support codeAction/resolve,
or when one of the servers does not support textDocument/definition
(like ruff)) we don't need to tell the user.