Option for logging to file instead of the *debug* buffer
Today, "kak-lsp --log somefile" still exists but presumably it's set
by users who don't know yet that the log already shows up in the
*debug* buffer.  To speed up adoption, the log is written to both
"somefile" and the *debug* buffer.

Of course the duplication is awkward, which is why "--log" is deprecated.

Unfortunately our logs today can add a lot of noise to the *debug*
buffer.  We should fix all of that but sometimes it's not so easy to
fix because it may depend on a language server.

Let's add a handbrake to redirect all log output ot a file.

Closes #782