Remove redundant errors when language server is not installed
When a language server for some filetype is not installed, we print a
few lines to the *debug* buffer for each new buffer, before disabling
LSP for that buffer.
Let's get rid of the redundant logs by keeping track of which buffers
we already disabled LSP in.

This is quite ugly because currently dont' have a way to remove the
tombstones.  This means that if a user deletes a buffer and recreates
it, LSP will remain disabled.

Since that only happens if a server is not installed,
this doesn't seem critical to fix.