Silence breadcrumbs error when client has switched to a different window
If I run lsp-references on a function with multiple references,
my client will switch to an lsp-goto buffer.
I frequently get this error

	error running command 'evaluate-commands -client client0
	-verbatim -- evaluate-commands -draft -client client0 --
	'buffer ''/home/johannes/git/kakoune-lsp/t/test.py'';
	set-option window lsp_modeline_breadcrumbs '' '''': 1:1:
	'evaluate-commands': 1:52: 'set-option': no window in context

Probably this is because we're both evaluating in a draft context and
switching buffer first. If the "buffer" command doesn't need to switch
buffer, everything is fine and we can use the calling window context.
But if we do, there is no window. I don't think there is a way to
access an inactive window.  Let's silence this error, it's harmless.