Ignore -v argument, to be replaced by --debug for now
This fixes a case where debug logging is accidentally left enabled.
This used to be harmless when logging only to a file
but now that we write to the debug buffer, it causes issues.

So let's just make "-vvv" stop working; users who
need to make a debugging excursion should be able to easily figure
out how to re-enable debug loggin, since we always log:

	LSP: INFO kak-lsp server starting. To control log verbosity, set the 'lsp_debug' option

We could probably just ignore "-vvv" in all cases but I left it for
now in case we need to bisect some issue.

Part of #782