Only block LSP commands that are run from hooks
The original idea of this blocking logic
(as can be seen pre-fb42037d (Stop clobbering the disabled_hooks option, 2024-09-27))
is to
1. prevent the debug buffer from spawning LSP requests and creating a feedback cycle
2. optimize away shell calls in all scratch buffers

Both cases are only relevant if lsp-send/lsp-did-open/lsp-did-change is
called from a hook. If not, we still want to take the slow path and get
a proper error and possibly a fallback command (see the parent commit).