README: link to troubleshooting section
Our readme is huge so make this more visible until we split it up.
While at it, rewrite it a little.

Remove the "crashes" section; it is good to know but now that we
have *debug*, I don't think any user needs to know this to debug
their issues.  I might be wrong, maybe the approach with a second
terminal is less scary to most users? Not sure.. but it is awkward,
so let's prefer to teach the non-awkward approach.