Elide temporary file when writing to fifo
We have received reports of slowness on Linux systems with non-tmpfs
/tmp and also on macOS. On  lsp-did-change invocations take >100ms.

I can reproduce a lot of slowness on an underpowered, virtualized
Ubuntu system.  On that system I ran

	mkfifo fifo && while cat fifo; do :; done

in one terminal and repeatedly run

	time echo | tee fifo

in another terminal. This occasionally (1/20 times) takes around 40ms.
Usually it's closer to 4ms.  On my non-virtualized Linux it never
takes more than 3ms.

With the currently available approaches we can't really work around
such OS/FS slowness, but we can at least avoid disk writes when
sending requests, like we used to do pre-18.0.0.

The "echo -to-file" was added mainly as part of the larger rework
to use more native Kakoune commands instead of shell expansion.
Initially the plan was to get rid of the shell entirely, and do
something similar to https://github.com/tomKPZ/pykak.  However,
since there are some scenarios where opening the fifo might hang,
and we don't want to risk data loss by freezing the entire Kakoune
session, we still wrap the fifo write into a cancellable shell process.

So let's get rid of the temp file and move this back into the shell
snippet.

Note that I tried to use "echo -to-shell-script" but that doesn't
really work because it offers no way of running "fail" commands that
are displayed in the status line. Not even command fifo seemed to work
IIRC but that we want to avoid as well because /tmp might not be tmpfs.