Remove redundant fifo re-creation
After we open the fifo, we do "rm fifo && mkfifo fifo" to create
a new fifo object.  Looks like this was a leftover from when I was
experimenting with sending LSP commands without any shell process
at all.  That didn't work out for other reasons (see the message in
parent commit and the link below).

The shell-less transport looked like 

	echo -to-file fifo ...

on the Kakoune side, and a reader loop on the kak-lsp server side.
Since it's always the same two OS processes reading/writing the fifo,
the kernel reuses the underlying object so the EOF signal is not
reliable, because a new write might have arrived or not.  That's why
we added the "rm fifo && mkfifo fifo" logic.  But since we already have
a shell process where such reuse won't happen, we can simply drop this.

In addition to the parent commit, this gets rid of disk access when
/tmp is not mounted as tmpfs.