Use rust-analyzer from PATH
I'm not sure if the original motivation for this is valid.  AFAICT,
rust-analyzer should be using the same mechanisms as rustc to resolve
the active toolchain version.  Simply using rust-analyzer from PATH
will lead to a simpler future (where it's easier to override for
users etc.).

I could be wrong about this but it matches other LSP clients so we
should be fine.

This happens to work around an issue where kak-lsp is continuously
restart if not installed.

Closes #823