Wait for PID file creation after start
By the time kak-lsp has created the PID file, it definitely
has created fifos, so we use this as synchronization point.

When running lsp-send twice in quick succession, both might start a
server because the second one won't wait for the first one to create
the PID file.  Let's fix that.