Refactor temporary directory cleanup logic
Since we want kak-lsp to handle things like SIGTERM, our cleanup ought
to comply with signal-safety(7).  The code to do this is a bit awkward.
Since it doesn't use destructors, it has some latent issues where it
leaves around the temporary directory if mkfifo() fails, for example.

Note that this fails to handle errors in rmdir() and unlink(), and
also fails to handle signals that come in before we register our
signal handler.  In future, we should at least fix signal handling