lsp-exit to wait until the session directory is removed
Similar to an earlier commit ("lsp-start to wait until existing server
has exited"), this fixes a case where LSP commands are used
too soon after lsp-exit. Also the case of "lsp-disable; lsp-enable".

After lsp-exit, we can't send LSP requests until the session directory
is gone.  What usually happens is that we get a "broken pipe" error,
or potentially a "no such file or directory".  This error is not
necessary, so let lsp-exit wait.  This should be a simpler behavior;
it allows to restart LSP via disable/enable mappings without errors.
It's a bit slow right now but the next commit will fix that.