Fix race conditions reading kak-lsp PID file
If lsp-send successfully reads "%%file{%opt{lsp_pid_file}}",
and then finds that the PID has changed,
then it reads the file again.

This definitely seems like a source of bugs.  Re-reading the PID
file here is only done for the case that kak-lsp was not running,
meaning we just restarted it here.
Let's move that special case to a different code path.

Also, make the write to the PID file atomic, so Kakoune will never
read an empty file and wrongly reset the lsp fifo twice.