Don't send SIGTERM to language server, remove obsolete wait()
It's not our job to kill a broken language server; commit 05880b9e
(don't hang if language server hung; ref #217, 2019-07-30) did that
as a workaround because it needed to wait() for the server to avoid
zombies.  Since commit de9a163c (Fix regression causing "kak -p" to
become a zombie, 2024-11-09), the wait() is useless (in fails with
an error) because we use SA_NOCLDWAIT.

I guess this approach might not be the best because we'll never notice
unclean shutdown of "kak -p" or the language server.  In future,
we should probably register a SIGCHLD handler and wait(WNOHANG)
in our event loop; logging any failures.