Work around hang due to lost fifo on macOS
Running test/clangd-invalid-utf8.sh for a thousand times on macOS
usually reproduces a hang.

Kakoune's "echo -to-file fifo1" succeeds but for some reason, kak-lsp's
"open(fifo1)" never returns.  Then Kakoune does "echo -to-file fifo2"
which hangs forever because we are out of sync.
I've attached a debugger and seen it hang like that in our KakEnd's
lsp-exit, and also inside NormalIdle's lsp-code-lens-request.

I think I tried switching kak-lsp's fifo reader from an unbounded queue
to one that only starts a new open() when the previous request has been
processed. That seems to make the hang less frequent but it's no fix.

I plan to reduce my reproducer; hopefully it's either our bug or we
can fix it.

For now let's wrap it in a shell process again, to work around
the hang.  Unfortunately the test probably still fails, unless one
installs the timeout (which has been added to POSIX).

So far I failed to reporoduce with

	mkfifo fifo{1,2}; cc test.c && WRITER=1 ./a.out & ./a.out

where test.c is

	#include <assert.h>
	#include <fcntl.h>
	#include <stdio.h>
	#include <stdlib.h>
	#include <unistd.h>
	int main() {
	  const char *fifos[2] = {"fifo1", "fifo2"};
	  int fd;
	  int fifo_index = 0;
	  while (1) {
	    open(fifos[fifo_index], getenv("WRITER") ? O_WRONLY : O_RDONLY);
	    fifo_index = 1 - fifo_index;
	    usleep(rand() % 1000);
	    assert(fd != -1);
	    close(fd);
	    putchar('.');
	    usleep(rand() % 1000);
	  }
	}