Fail startup if session state files already exists
In this sequence

	kak -s foo -e 'eval %sh{kak-lsp}; lsp-enable'
	rm /run/user/$EUID/kakoune-lsp/foo.ref
	pkill -9 kak
	kak -clear
	kak -s foo -e 'eval %sh{kak-lsp}; lsp-enable'

The last command tries to reuse the kak-lsp session.  It should not,
since the old session directory at /run/user/$EUID/kakoune-lsp/foo
was not properly cleaned up, so this is not really safe. Make this
a hard error.

Only allow this if kak-lsp was started outside the editor,
for debugging and backwards compatibility.