Stop linking to the Wiki page for installing servers
This page is somewhat spotty and suggests things that may not
be best practice.  Let's instead point to our source of truth in
rc/servers.kak (which is maintained better), and leave the rest to
user's package manager.

While at it, remove the stanza on versioning, which hasn't been
relevant in a while.