Bravely remove obsolete command_fifo, make response_fifo handling more robust
Commit b54ec807 (Allow *-sync requests if the controller is down,
2022-03-27) made it so that inside *-sync requests
we send commands to $kak_command_fifo instead of "kak -p".

This is because back then "kak -p" would deadlock when the parent
LSP request would restart kak-lsp.  That commit claimed that starting
kak-lsp "requires some roundtrips to the editor", hence the deadlocked
with "kak -p".

Since 5b787ccc (Simplify sending of initialization options, 2024-11-10)
we no longer do a roundtrip to the editor when restarting.

This means that when answering, say lsp-formatting-sync,
the things we want to send to the editor (besides the formatting response) are: 
1. logs for the *debug* buffer
2. things like diagnostics (in response to notifications from the server)

I guess both can wait until after lsp-formatting-sync is done.

Since we currently don't wait for "kak -p" to return, we can simply
ignore those.

Let's assert that we don't need the full power of command_fifo.  Also,
let's make the response_fifo handling much less brittle by using RAII
to make sure it's written to exactly once.

The regression here is that a request like lsp-formatting-sync that
needs to start the server will cause the "language server still
not initialized" error to be printed.  Let's add a special case to
suppress this message, since it's pretty much never interesting in
the sync case.