Fix lsp-did-change not being sent for lsp-code-actions-sync
Also we can extract this now.

This should allow us to assert that synchronous (fifo-bearing)
requests are never stale.

This stale detection was added in c622f62d (Forward requests only
when they have the correct buffer version, 2023-01-28)
when buffer synchronization was racy;
even a

	lsp-did-change
	lsp-foo

was not guaranteed to arrive in that order because each spawned its
own "kak-lsp --request" subprocess and not wait for it.
That is now fixed, since we have FIFO semantics.

Nevertheless it looks like we want to keep staleness detection,
because for non-fifo requests we rely on NormalIdle and InsertIdle
to send textDocument/didChange.