Rework editor-command sending
Commit 05db0d26 (Remove shell calls from async request sending,
2024-11-05) removed the wait() call from "kak -p" to avoid a
deadlock. The deadlock was possible because we used blocking fifos at
that time.  Now that we use nonblocking fifos, most deadlocks can't
happen even if we wait on "kak -p".

There is one potential deadlock case left though, namely the *-sync
family of commands. While those are running, we can't close any
"kak -p" process.

The problem with not waiting is that concurrent "kak -p" processes
will race, causing nondeterministic ordering of log statements. Let's
add the wait back. We avoid the aforementioned deadlock by running
these processes on a background thread. (Would be nice to get rid of
that eventually.)

Take this opportunity to get rid of SIGCHLD handling.  As mentioned
in 20b2dcc8 (Don't block when language servers are slow to exit,
2024-11-15), it's fine to do a potentially slow wait() on language
server shutdown, because at this point the kak-lsp server has already
detached from its Kakoune session (and another one may already have
started).

Also move the computation of the input for "kak -p" to before we start
said process.  Given that this computation might recursively spawn a
"kak -p" process for logging, it seems simpler to do it this way,
to never have more than one such process alive at a time.

To facilitate this, consolidate how we send commands to the editor.
Some places start "kak -p" from the calling thread instead of the
"Messages to editor" thread.  Get rid of all of those cases (except
when there is no "Messages to editor" thread, or it's the current
thread).

Plus some related cleanup and fixes.

Fixes #812