Make sure to call textDocument/didOpen when lsp_servers is set after BufCreate
This fixes an assertion failure in ensure_did_open() in the scenario
where a user defines their own lsp_servers hooks (after our 
"BufCreate .* %{ lsp-did-open } " hook).

This is quite brittle; let's try to cover another crash by using
-always; though we should get rid of that eventually.

We might want to go back to putting the server side in charge of this
bookkeeping. But that will require roundtrips (with more potential for
error) or reading from disk which is not accurate but okay in practice.