Config knob to override magic single-instance setting
As of commit fe4b4810 (Reuse language server across projects if it
supports workspaceFolders, 2024-09-04) we use a single instance of
rust-analyzer for multiple projects.

This is done if the server supports
`workspaceFolders`. This is an [acknowledged
hack](https://github.com/rust-lang/rust-analyzer/issues/18270#issuecomment-2403331913),
working around the issue that LSP does not provide an
way to determine whether a server should be reused, see
https://github.com/microsoft/language-server-protocol/issues/674.

`workspaceFolders` is somewhat close; it indicates that the server
supports multiple roots.

Besides being able to detect project roots, kakoune-lsp has no
notion of projects. Nor does Kakoune. There is no project scope,
and no configuration section for projects. This may change in future
but hopefully we can get by with a simpler approach (i.e. simply set
the relevant options in all buffers pertaining to a project).

Hence our "active" project is determined based on %val{buffile},
%sh{pwd}, and the root and root_globs options.  Notably there is no
plural -- we do not support multiple simultaneous active projects at
this point.

This was the main reason for interpreting `workspaceFolders` to mean "single
instance": if the server supports multiple roots ( while we definitely
do not), just assume it can handle all roots in this Kakoune session.
We do not dynamically add any root to workspaceFolders, so the
server only uses one -- the one it was started with.  This seems to
be as good as it gets under the current limitations of our incomplete
multi-root support.

Unfortunately single-instance is not always the correct
thing to do. Since single-instance is unlike historical
behavior, it definitely breaks lots of scenarios, for
example when using different server-specific configuration
in different projects using lua-language-server, see
https://github.com/kakoune-lsp/kakoune-lsp/pull/765#issuecomment-2626156962

Let's add a knob to override it. Set it explicitly for rust-analyzer
even though that's redundant.