controller: consume entire request before checking for user errors
As reported in
https://github.com/kakoune-lsp/kakoune-lsp/commit/a7b6f50bbbfb779315a88ee2da2279cc27cf5e4d#r156588552,
a buggy config like

	set-option buffer lsp_servers %{
		[language.c]
		filetypes = [ "c", "cpp" ]
		roots = [ ".git" ]
		command = "clangd"
	}

causes a crash.  Before crashing, we correctly report an error but fail to
consume the invalid request in its entirety.  As a result, the next request
is corrupted, causing a crash.

Fix this by always consuming the entire input before checking for user errors.