Transpose the roots and `cd ..` loops when finding the project root
A bit subjective, but to me when the project is looking for roots it
should stop at the first directory matching any of the globs, rather
than going through each glob in order and matching the first directory
containing it.

It leads to counterintuitive stuff like a project that uses `setup.py`
and `requirements.txt` isn't recognized as the root because a parent
contains a `pyproject.toml` as a uv workspace.