Fix ctx.exec() API, remove redundant "evaluate-commands"
The command argument to ctx.exec() must be a single Kakoune command,
i.e. arbitrary Kakoune scripts like "nop; nop" are not allowed.

This API limitation doesn't make sense, and we work around it
in a few places by wrapping multi-line commands into another
"evaluate-commands", (which implies another round of quoting)

The reason for this limitation (avoid many nested quotes) is not very
strong and we can achieve it in a different way. For now let's use
a conservative heuristic for that.

In future, callers could pass explicit objects like enum variants
EditorCommand::Single or EditorCommand::Script.