Add [server_name] prefix to code action titles
If multiple servers provide code actions, the code action menu can
be chaotic.  Let's add the server name as prefix in this case (like we
do for lsp-diagnostics), to help users orient themselves and filter
through the list faster (both because it'll be sorted alphabetically
and because they can add a mapping that adds $server_name to the
command prompt).

Closes #847