Fix lsp-find-error incorrectly handling --previous flag
In the lsp-find-error command, the $previous variable was only set when --previous was passed, and remained unset otherwise. Evaluating an empty string as a boolean with if statements is always true. Explicitly setting previous=false as the default state, and overriding to previous=true if the --previous flag was given fixes the issue.